adding utils files split from main
